refactor: merge tag_database.py into tag_search.py
