+++ src/tagstudio/qt/ts_qt.py
@@ -86,8 +86,7 @@ from tagstudio.qt.mixed.migration_modal import JsonMigrationModal
 from tagstudio.qt.mixed.progress_bar import ProgressWidget
 from tagstudio.qt.mixed.settings_panel import SettingsPanel
 from tagstudio.qt.mixed.tag_color_manager import TagColorManager
-from tagstudio.qt.mixed.tag_database import TagDatabasePanel
-from tagstudio.qt.mixed.tag_search import TagSearchModal
+from tagstudio.qt.mixed.tag_search import TagSearchModal, TagSearchPanel
 from tagstudio.qt.models.palette import ColorType, UiColor, get_ui_color
 from tagstudio.qt.platform_strings import trash_term
 from tagstudio.qt.previews.vendored.ffmpeg import FFMPEG_CMD, FFPROBE_CMD
@@ -364,7 +363,7 @@ class QtDriver(DriverMixin, QObject):
 
         # Initialize the Tag Manager panel
         self.tag_manager_panel = PanelModal(
-            widget=TagDatabasePanel(self, self.lib),
+            widget=TagSearchPanel(self.lib, is_tag_chooser=False, driver=self),
             title=Translations["tag_manager.title"],
             done_callback=lambda checked=False: (
                 self.main_window.preview_panel.set_selection(self.selected, update_preview=False)
